requirements.txt
fastapi
uvicorn[standard]
openai-whisper
torch
soundfile
python-multipart